+++ src/rcpp_rrap_squared_distance.cpp
@@ -20,22 +20,21 @@ std::vector<double> rcpp_rrap_squared_distance(Rcpp::NumericMatrix pu_coordinate
   //// Preliminary processing
   // declare variables
   std::size_t N_pu_INT = pu_coordinates_MTX.rows();
-  std::size_t N_dp_INT = dp_coordinates_MTX.rows(); 
-  std::size_t N_k_INT = pu_coordinates_MTX.cols(); 
+  std::size_t N_dp_INT = dp_coordinates_MTX.rows();
   double currProb, curr_value;
 
   //// Main processing
   std::vector<double> d(N_dp_INT);
   Eigen::ArrayXd tmp_AXD(N_pu_INT);
   std::vector<size_t> pu_ids(N_pu_INT);
-  
+
   for (std::size_t j=0; j < N_dp_INT; ++j) {
     // init
     std::iota(pu_ids.begin(), pu_ids.end(), 0);
-    
+
     // calculate distances
     tmp_AXD = (pu_coordinates_MTX.rowwise() - dp_coordinates_MTX.row(j)).rowwise().squaredNorm().transpose();
-    
+
     // sort pus in order of distance
     std::partial_sort(
       pu_ids.begin(), pu_ids.begin()+maximum_r_level, pu_ids.end(),
@@ -43,7 +42,7 @@ std::vector<double> rcpp_rrap_squared_distance(Rcpp::NumericMatrix pu_coordinate
         return(tmp_AXD[p1] < tmp_AXD[p2]);
       }
     );
-    
+
     // calculate expected weighted distances for real pus
     currProb=1.0;
     curr_value=0.0;
@@ -53,13 +52,12 @@ std::vector<double> rcpp_rrap_squared_distance(Rcpp::NumericMatrix pu_coordinate
     }
     // calculate expected weighted distance for failure PU
     curr_value+=(currProb*failure_distance);
-    
+
     // update value
     d[j] = curr_value;
   }
-  
+
   //// Exports
   // return proportion held
   return(d);
 }
-